- Fixing bugs in the main dialog when "draw in new tab" windows, where the old values are overwrote - add the user interface part of the saving canvases definition to file (see File -> Import/Export Canvases in the main dialog)

+++ Canvas.cpp
@@ -64,7 +64,7 @@ void Canvas::setGraphType (graphTypes graphType)
 void Canvas::streamToFile(ofstream & s)
 {
     // output header in format: [Canvas:_type]
-    s << "[Canvas:"; s << type; s << "]\n";
+    s << "[Canvas:"; s << (int) type; s << "]\n";
     s << "Name = " << getName() << "\n";
     s << "EventCut = " << getEventCut() << endl;
     
@@ -141,5 +141,3 @@ void Canvas::print()
 {
     getGraphInfo()->print();
 }
-
-
